# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rewaa Quran Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginInput {
  /** Email or phone number */
  identifier: string;
  password: string;
}

export type AuthResponseRole = typeof AuthResponseRole[keyof typeof AuthResponseRole];


export const AuthResponseRole = {
  student: 'student',
  teacher: 'teacher',
  admin: 'admin',
} as const;

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  student: 'student',
  teacher: 'teacher',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  role: UserProfileRole;
  /** @nullable */
  age?: number | null;
  createdAt?: string;
}

export interface AuthResponse {
  user: UserProfile;
  role: AuthResponseRole;
}

export interface Student {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  level?: string | null;
  /** @nullable */
  teacherId?: number | null;
  /** @nullable */
  teacherName?: string | null;
  role: string;
  createdAt?: string;
}

export interface StudentInput {
  name: string;
  email?: string;
  phone?: string;
  password: string;
  age?: number;
  level?: string;
  /** @nullable */
  teacherId?: number | null;
}

export interface StudentUpdate {
  name?: string;
  email?: string;
  phone?: string;
  age?: number;
  level?: string;
  /** @nullable */
  teacherId?: number | null;
}

export interface Teacher {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  role: string;
  /** @nullable */
  specialization?: string | null;
  studentCount?: number;
  createdAt?: string;
}

export interface TeacherInput {
  name: string;
  email?: string;
  phone?: string;
  password: string;
  specialization?: string;
}

export interface TeacherUpdate {
  name?: string;
  email?: string;
  phone?: string;
  specialization?: string;
}

export type ClassStatus = typeof ClassStatus[keyof typeof ClassStatus];


export const ClassStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Class {
  id: number;
  dayOfWeek: string;
  time: string;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  teacherId: number;
  /** @nullable */
  teacherName?: string | null;
  isOnline?: boolean;
  /** @nullable */
  meetingLink?: string | null;
  status?: ClassStatus;
  createdAt?: string;
}

export type ClassInputStatus = typeof ClassInputStatus[keyof typeof ClassInputStatus];


export const ClassInputStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface ClassInput {
  dayOfWeek: string;
  time: string;
  studentId: number;
  teacherId: number;
  isOnline?: boolean;
  /** @nullable */
  meetingLink?: string | null;
  status?: ClassInputStatus;
}

export type ClassUpdateStatus = typeof ClassUpdateStatus[keyof typeof ClassUpdateStatus];


export const ClassUpdateStatus = {
  scheduled: 'scheduled',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface ClassUpdate {
  dayOfWeek?: string;
  time?: string;
  studentId?: number;
  teacherId?: number;
  isOnline?: boolean;
  /** @nullable */
  meetingLink?: string | null;
  status?: ClassUpdateStatus;
}

export type AchievementGrade = typeof AchievementGrade[keyof typeof AchievementGrade];


export const AchievementGrade = {
  excellent: 'excellent',
  good: 'good',
  acceptable: 'acceptable',
  needs_improvement: 'needs_improvement',
} as const;

export interface Achievement {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  teacherId?: number | null;
  /** @nullable */
  teacherName?: string | null;
  surah: string;
  startAyah: number;
  endAyah: number;
  grade: AchievementGrade;
  /** @nullable */
  notes?: string | null;
  date: string;
  createdAt?: string;
}

export type AchievementInputGrade = typeof AchievementInputGrade[keyof typeof AchievementInputGrade];


export const AchievementInputGrade = {
  excellent: 'excellent',
  good: 'good',
  acceptable: 'acceptable',
  needs_improvement: 'needs_improvement',
} as const;

export interface AchievementInput {
  studentId: number;
  surah: string;
  startAyah: number;
  endAyah: number;
  grade: AchievementInputGrade;
  notes?: string;
  date: string;
}

export type AchievementUpdateGrade = typeof AchievementUpdateGrade[keyof typeof AchievementUpdateGrade];


export const AchievementUpdateGrade = {
  excellent: 'excellent',
  good: 'good',
  acceptable: 'acceptable',
  needs_improvement: 'needs_improvement',
} as const;

export interface AchievementUpdate {
  surah?: string;
  startAyah?: number;
  endAyah?: number;
  grade?: AchievementUpdateGrade;
  /** @nullable */
  notes?: string | null;
  date?: string;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  class_reminder: 'class_reminder',
  achievement_added: 'achievement_added',
  schedule_change: 'schedule_change',
  general: 'general',
} as const;

export interface Notification {
  id: number;
  userId: number;
  title: string;
  message: string;
  isRead: boolean;
  type?: NotificationType;
  createdAt: string;
}

export interface Attendance {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  classId: number;
  date: string;
  present: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt?: string;
}

export interface AttendanceInput {
  studentId: number;
  classId: number;
  date: string;
  present: boolean;
  notes?: string;
}

export type DashboardStatsAchievementsByGradeItem = {
  grade: string;
  count: number;
};

export interface DashboardStats {
  totalStudents: number;
  totalTeachers: number;
  totalClasses: number;
  totalAchievements: number;
  classesToday: number;
  classesThisWeek: number;
  achievementsByGrade?: DashboardStatsAchievementsByGradeItem[];
  recentAchievements?: Achievement[];
}

export interface StudentStats {
  studentId: number;
  totalAchievements: number;
  totalClasses: number;
  attendedClasses: number;
  completionRate: number;
  lastAchievement?: Achievement;
  upcomingClasses?: Class[];
}

export type ListAchievementsParams = {
studentId?: number;
};

export type ListAttendanceParams = {
studentId?: number;
classId?: number;
};