# lib/api-zod/src/generated/types/achievementUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rewaa Quran Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { AchievementUpdateGrade } from './achievementUpdateGrade';

export interface AchievementUpdate {
  surah?: string;
  startAyah?: number;
  endAyah?: number;
  grade?: AchievementUpdateGrade;
  /** @nullable */
  notes?: string | null;
  date?: Date;
}