# lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rewaa Quran Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Attendance {
  id: number;
  studentId: number;
  /** @nullable */
  studentName?: string | null;
  classId: number;
  date: Date;
  present: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt?: Date;
}